Move application_path to constants
