+++ src/application/a_knowledge.cpp
@@ -4,8 +4,6 @@ namespace Knowledge
 {
 UniquePtr<Player> Daddy_Player;
 
-std::string application_path;
-
 Constants::State::W curr_state = Constants::State::W::Home;
 
 /// Global variables but available only 
@@ -43,28 +41,6 @@ sf::Event GetEvent()
 void SetEvent(sf::Event ev)
 {
     event_ = ev;
-}
-
-void CreateApplicationPath()
-{
-    Logger::Get() << "Creating applicaition folders\n";
-
-    system("mkdir $HOME/.musicalbash -p");
-    system("mkdir $HOME/.musicalbash/downloads -p");
-
-    char path[1024];
-    path[0] = 0;
-    
-    FILE *f = popen("(cd $HOME/.musicalbash; pwd)", "r");
-    fgets(path, 1024, f);
-
-    if (path[0] == 0)
-        throw bad_behaviour("Unable to find application path!");
-    
-    application_path = std::string(path);
-    application_path.pop_back();
-
-    Logger::Get() << "Found app path at \"" + application_path + "\"\n";
 }
 
 }
